skills/good-mp-post/requirements.txt
# WeChat API (existing)
requests>=2.28.0
python-dotenv>=1.0.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.35

# Markdown processing
markdown==3.5.1
